% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyhtml_document.R
\name{html_dependency_lazyload}
\alias{html_dependency_lazyload}
\title{Provide HTML dependencies for lazyhtml R Markdown format}
\usage{
html_dependency_lazyload()
}
\description{
Provide HTML dependencies for lazyhtml R Markdown format
}

