% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.htmlwidget}
\alias{knit_print.htmlwidget}
\title{A custom knitr print method for htmlwidgets}
\usage{
\method{knit_print}{htmlwidget}(x, ..., options = NULL)
}
\arguments{
\item{x}{an R object to be printed}

\item{options}{knitr options}

\item{\ldots}{additional arguments passed to the htmlwidgets \code{knit_print} method}
}
\description{
A custom knitr print method for htmlwidgets
}

