% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyhtml_document.R
\name{lazy_render}
\alias{lazy_render}
\title{Format for converting from R Markdown to an HTML document with lazy loading of graphics}
\usage{
lazy_render(..., rmarkdown = "html_document")
}
\arguments{
\item{rmarkdown}{rmarkdown function to use to compile the document.  Defaults to \code{\link[rmarkdown]{html_document}}}

\item{\ldots}{parameters passed to rmarkdown's function matching the parameter \code{rmarkdown}}
}
\description{
Format for converting from R Markdown to an HTML document with lazy loading of graphics
}
\examples{
if (interactive()) browseURL("https://github.com/hafen/lazyrmd#usage")
}

