% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyhtml_document.R
\name{lazyhtml_document}
\alias{lazyhtml_document}
\title{Format for converting from R Markdown to an HTML document with lazy loading of graphics}
\usage{
lazyhtml_document(...)
}
\arguments{
\item{\ldots}{parameters passed to \code{\link[rmarkdown]{html_document}}}
}
\description{
Format for converting from R Markdown to an HTML document with lazy loading of graphics
}
\examples{
if (interactive()) browseURL("https://github.com/hafen/lazyrmd#usage")
}

