% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyhtml_document.R
\name{lazy_render}
\alias{lazy_render}
\title{Format for converting from R Markdown to an HTML document with lazy loading of graphics}
\usage{
lazy_render(..., lazyrmd_render_fn = "html_document",
  lazyrmd_render_package = "rmarkdown")
}
\arguments{
\item{lazyrmd_render_fn}{function used to compile the document.  Paired with the \code{lazyrmd_render_package} parameter, it defaults to \code{\link[rmarkdown]{html_document}}.}

\item{lazyrmd_render_package}{package used to look for the \code{lazyrmd_render_fn}.}

\item{\ldots}{parameters passed to rmarkdown's function matching the parameter \code{rmarkdown}}
}
\description{
Format for converting from R Markdown to an HTML document with lazy loading of graphics
}
\examples{
if (interactive()) browseURL("https://github.com/hafen/lazyrmd#usage")
}

